requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncmy
aiomysql
python-dotenv
